// 03-gifs-app/src/gifs/actions/get-gifs-by-query.action.ts
import type { GiphyResponse } from '../interfaces/giphy.response'
import type { Gif } from '../interfaces/gif.interface'
import { giphyApi } from '../api/giphy.api'

export const getGifsByQuery = async (query: string): Promise<Gif[]> => {
    //fetch(`https://api.giphy.com/v1/gifs/search?api_key=1TdRc3tqzUOE8qYWTOIJ6EQ2j3zcYJpC&q=${query}&limit=10&lang=es`)

    if (query.trim().length === 0) {
        return []
    }

    try {

        const response = await giphyApi<GiphyResponse>('/search', {
            params: {
                q: query,
                limit: 10,
            }
        })

        // const response = await axios.get<GiphyResponse>('/search', {
        //     params: {
        //         q: query,
        //         limit: 10,
        //         lang: 'es',
        //         api_key: import.meta.env.VITE_GIPHY_API_KEY,
        //         //api_key: '1TdRc3tqzUOE8qYWTOIJ6EQ2j3zcYJpC',
        //     }
        // })

        return response.data.data.map((gif) => ({
            id: gif.id,
            title: gif.title,
            url: gif.images.original.url,
            width: Number(gif.images.original.width),
            height: Number(gif.images.original.height),
        }))

    } catch (error) {
        console.error(error)
        return []
    }
}